Rewrote the website to React
I think we got 100% of the functionality rewritten! Thanks to react we can use Routing instead of classes to achieve seamless transitions, which is great. Also added page titles to differentiate between pages (before there were just blank pages and it was tough to differentiate between anything that was not the contacts page)
TODO: 404 Page